move cron folder to /scorekeeper/jobs/cron
